refactor api result handling
