+++ shared/src/commonMain/kotlin/com/bunbeauty/shared/data/di/NetworkModule.kt
@@ -2,8 +2,6 @@ package com.bunbeauty.shared.data.di
 
 import com.bunbeauty.shared.Logger
 import com.bunbeauty.shared.Logger.NETWORK_TAG
-import com.bunbeauty.shared.data.network.api_result_handler.ApiResultHandler
-import com.bunbeauty.shared.data.network.api_result_handler.ApiResultHandlerDelegate
 import com.bunbeauty.shared.httpClientEngine
 import io.ktor.client.HttpClient
 import io.ktor.client.plugins.DefaultRequest
@@ -61,8 +59,5 @@ fun networkModule() = module {
                 }
             }
         }
-    }
-    factory<ApiResultHandler> {
-        ApiResultHandlerDelegate()
     }
 }
