+++ shared/src/commonMain/kotlin/com/bunbeauty/shared/data/di/RepositoryModule.kt
@@ -54,7 +54,6 @@ fun repositoryModule() = module {
             orderDao = get(),
             networkConnector = get(),
             orderMapper = get(),
-            apiResultHandler = get(),
         )
     }
     single<MenuProductRepo> {
@@ -119,14 +118,6 @@ fun repositoryModule() = module {
         )
     }
     single {
-        PaymentRepository(
-            networkConnector = get(),
-            paymentMethodMapper = get(),
-            paymentMethodDao = get(),
-            apiResultHandler = get(),
-        )
-    }
-    single {
         SettingsRepository(
             dataStoreRepo = get(),
             networkConnector = get(),
@@ -138,7 +129,6 @@ fun repositoryModule() = module {
             networkConnector = get(),
             paymentMethodMapper = get(),
             paymentMethodDao = get(),
-            apiResultHandler = get(),
         )
     }
 }
