+++ shared/src/commonMain/kotlin/com/bunbeauty/shared/data/repository/PaymentRepository.kt
@@ -3,18 +3,15 @@ package com.bunbeauty.shared.data.repository
 import com.bunbeauty.shared.data.dao.payment_method.IPaymentMethodDao
 import com.bunbeauty.shared.data.mapper.payment.PaymentMethodMapper
 import com.bunbeauty.shared.data.network.api.NetworkConnector
-import com.bunbeauty.shared.data.network.api_result_handler.ApiResultHandler
-import com.bunbeauty.shared.domain.model.Payment
 import com.bunbeauty.shared.domain.model.payment_method.PaymentMethod
 import com.bunbeauty.shared.domain.repo.PaymentRepo
+import com.bunbeauty.shared.extension.getNullableResult
 
 class PaymentRepository(
     private val networkConnector: NetworkConnector,
     private val paymentMethodMapper: PaymentMethodMapper,
     private val paymentMethodDao: IPaymentMethodDao,
-    apiResultHandler: ApiResultHandler
-) : PaymentRepo,
-    ApiResultHandler by apiResultHandler {
+) : PaymentRepo {
 
     private var paymentMethodListCache: List<PaymentMethod>? = null
